Move parse up a level, it depends on ast and parser
