+++ sdk/burdock/sql/__init__.py
@@ -1,9 +1,8 @@
 from burdock.metadata.collection import CollectionMetadata
-from .ast.parse import QueryParser
-from .ast.validate import Validate
 from .private_rewrite import Rewriter
 from .private_reader import PrivateReader
 from burdock.reader.sql.pandas import PandasReader
+from .parse import QueryParser
 
 
 def execute_private_query(reader, schema, budget, query):
@@ -11,5 +10,5 @@ def execute_private_query(reader, schema, budget, query):
     query = reader._sanitize_query(query) if hasattr(reader ,"_sanitize_query") else query
     return PrivateReader(reader, schema, budget).execute(query)
 
-__all__ = ["CollectionMetadata", "QueryParser", "Rewriter", "Validate",
+__all__ = ["CollectionMetadata", "Rewriter", "QueryParser",
            "PandasReader", "execute_private_query"]
