+++ tests/query/test_interval_expr.py
@@ -1,10 +1,8 @@
 import pytest
 import pandas as pd
 
-from burdock.sql import PandasReader
-from burdock.sql import CollectionMetadata
-from burdock.sql import QueryParser
-from burdock.sql.private_reader import PrivateReader
+from burdock.sql import PandasReader, CollectionMetadata, PrivateReader
+from burdock.sql.parse import QueryParser
 from burdock.reader.sql.rowset import TypedRowset
 from pandasql import sqldf
