+++ tests/query/test_types.py
@@ -1,6 +1,7 @@
 from os.path import dirname, join
 
-from burdock.sql import CollectionMetadata, QueryParser
+from burdock.sql import CollectionMetadata
+from burdock.sql.parse import QueryParser
 
 dir_name = dirname(__file__)
